FIX - compile after system update
* There are always these issues with cstdint. Sometimes it is
  included already, sometimes not....

+++ tof/dataclasses/C++/include/packets/RBEnvPacket.h
@@ -6,63 +6,64 @@
 #include <ostream>
 
 #include "RBEnvData.h"
+#include "tof_typedefs.h"
 
 // size of the packet is fixed, in bytes
 // 11 fields a 2 bytes, + 2 * 2 bytes header and tail each + 1 byte rbid
-static const uint8_t RBENVPACKETSIZE = 27;
+static const u8 RBENVPACKETSIZE = 27;
 
 // sensor ranges
-static const float TEMPSENSMAX = 100;
-static const float TEMPSENSMIN = 0;
+static const i32 TEMPSENSMAX = 100;
+static const i32 TEMPSENSMIN = 0;
 
-static const float VOLTSENSMAX = 100;
-static const float VOLTSENSMIN = 0;
+static const i32 VOLTSENSMAX = 100;
+static const i32 VOLTSENSMIN = 0;
 
-static const float CURRENTSENSMAX = 100;
-static const float CURRENTSENSMIN = 0;
+static const i32 CURRENTSENSMAX = 100;
+static const i32 CURRENTSENSMIN = 0;
 
-static const float POWERSENSMAX = 100;
-static const float POWERSENSMIN = 0;
+static const i32 POWERSENSMAX = 100;
+static const i32 POWERSENSMIN = 0;
 
-static const float PREAMPTEMPSENSMAX = 100;
-static const float PREAMPTEMPSENSMIN = 0;
+static const i32 PREAMPTEMPSENSMAX = 100;
+static const i32 PREAMPTEMPSENSMIN = 0;
 
-static const float PREAMPBIASSENSMAX = 100;
-static const float PREAMPBIASSENSMIN = 0;
+static const i32 PREAMPBIASSENSMAX = 100;
+static const i32 PREAMPBIASSENSMIN = 0;
 
-static const float TEMPRBSENSMAX = 100; 
-static const float TEMPRBSENSMIN = 0;
+static const i32 TEMPRBSENSMAX = 100; 
+static const i32 TEMPRBSENSMIN = 0;
 
-static const float VOLTRBSENSMAX = 100; 
-static const float VOLTRBSENSMIN = 0;
+static const i32 VOLTRBSENSMAX = 100; 
+static const i32 VOLTRBSENSMIN = 0;
 
-static const float CURRENTRBSENSMAX = 100;
-static const float CURRENTRBSENSMIN = 0;
+static const i32 CURRENTRBSENSMAX = 100;
+static const i32 CURRENTRBSENSMIN = 0;
 
-static const float POWERRBSENSMAX = 100;
-static const float POWERRBSENSMIN = 0;
+static const i32 POWERRBSENSMAX = 100;
+static const i32 POWERRBSENSMIN = 0;
 
 struct RBEnvPacket {
 
  public:
-  uint16_t head = 0xAAAA;
-  uint16_t tail = 0x5555;
+  u16 head = 0xAAAA;
+  u16 tail = 0x5555;
 
   // readout board id
-  uint8_t rb_id;
+  u8 rb_id;
 
   // 12 bit sensors    
-  uint16_t temperature;
-  uint16_t voltage;
-  uint16_t current;
-  uint16_t power;
-  uint16_t preamp_temp;
-  uint16_t preamp_bias;
-  uint16_t temperature_rb;
-  uint16_t voltage_rb;
-  uint16_t current_rb;
-  uint16_t power_rb;
-  uint16_t lol_status;
+  u16 temperature;
+  u16 voltage;
+  u16 current;
+  u16 power;
+  u16 preamp_temp;
+  u16 preamp_bias;
+  u16 temperature_rb;
+  u16 voltage_rb;
+  u16 current_rb;
+  u16 power_rb;
+  u16 lol_status;
  
  public:
   void fill_from_envdata(RBEnvData* env_data);
@@ -100,32 +101,32 @@ struct RBEnvPacket {
 
 
   // getters
-  uint8_t get_rb_id(); 
-  float get_temperature();
-  float get_voltage();
-  float get_current();
-  float get_power();
-  float get_preamp_temp();
-  float get_preamp_bias();
-  float get_temperature_rb();
-  float get_voltage_rb();
-  float get_current_rb();
-  float get_power_rb();
-  float get_lol_status();
+  u8 get_rb_id(); 
+  i32 get_temperature();
+  i32 get_voltage();
+  i32 get_current();
+  i32 get_power();
+  i32 get_preamp_temp();
+  i32 get_preamp_bias();
+  i32 get_temperature_rb();
+  i32 get_voltage_rb();
+  i32 get_current_rb();
+  i32 get_power_rb();
+  i32 get_lol_status();
 
   // setters
-  void set_rb_id         (uint8_t rb_id); 
-  void set_temperature   (float value);
-  void set_voltage       (float value);
-  void set_current       (float value);
-  void set_power         (float value);
-  void set_preamp_temp   (float value);
-  void set_preamp_bias   (float value);
-  void set_temperature_rb(float value);
-  void set_voltage_rb    (float value);
-  void set_current_rb    (float value);
-  void set_power_rb      (float value);
-  void set_lol_status    (float value);
+  void set_rb_id         (u8 rb_id); 
+  void set_temperature   (i32 value);
+  void set_voltage       (i32 value);
+  void set_current       (i32 value);
+  void set_power         (i32 value);
+  void set_preamp_temp   (i32 value);
+  void set_preamp_bias   (i32 value);
+  void set_temperature_rb(i32 value);
+  void set_voltage_rb    (i32 value);
+  void set_current_rb    (i32 value);
+  void set_power_rb      (i32 value);
+  void set_lol_status    (i32 value);
 
 };
